Update bash scripts.

+++ gitop.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
+if [ -n "$1" ]
+then
 rm *.gch 
 rm a.out
 git add .
-git commit -m "$1"
+git commit -m "$*"
 git pull
 git push
+else
+echo "Enter commit message. "
+fi
